# APP/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/realtime/events": {
      "get": {
        "operationId": "RealtimeController_pullEvents",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Realtime"
        ]
      }
    },
...
    "/api/groups/{groupId}/messages": {
      "get": {
        "operationId": "MessagesController_listMessages",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "MessagesController_createMessage",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/groups/{groupId}/messages/{messageId}": {
      "patch": {
        "operationId": "MessagesController_editMessage",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/groups/{groupId}/messages/{messageId}/revoke": {
      "post": {
        "operationId": "MessagesController_revokeMessage",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
...
    "/api/users/{userId}/role": {
      "patch": {
        "operationId": "UsersController_updateUserRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{userId}": {
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{userId}/disabled": {
      "patch": {
        "operationId": "UsersController_setUserDisabled",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{userId}/password": {
      "patch": {
        "operationId": "UsersController_resetUserPassword",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "SekerChat HTTP API",
    "description": "Published contract for browser, CLI, reminder, and mobile clients.",
    "version": "0.8.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "seker_access"
      }
    },
    "schemas": {
...
      "MessageSenderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "avatarUrl"
        ]
      },
      "MessageAttachmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fileId": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "contentUrl": {
            "type": "string"
          },
          "metadataUrl": {
            "type": "string"
          },
          "uploaderId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "file"
            ]
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "isSharing": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "fileId",
          "groupId",
          "originalName",
          "mimeType",
          "size",
          "createdAt",
          "contentUrl",
          "metadataUrl",
          "uploaderId",
          "kind",
          "thumbnailUrl"
        ]
      },
      "MessageReplyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "file",
              "system"
            ]
          },
          "textPreview": {
            "type": "string",
            "nullable": true
          },
          "sender": {
            "$ref": "#/components/schemas/MessageSenderResponseDto"
          },
          "attachment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageAttachmentResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "senderId",
          "type",
          "textPreview",
          "sender",
          "attachment"
        ]
      },
      "MessageReadReceiptMemberResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "userId",
          "email",
          "displayName",
          "avatarUrl"
        ]
      },
      "MessageReadReceiptResponseDto": {
        "type": "object",
        "properties": {
          "totalRecipients": {
            "type": "number"
          },
          "readCount": {
            "type": "number"
          },
          "unreadCount": {
            "type": "number"
          },
          "readBy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageReadReceiptMemberResponseDto"
            }
          },
          "unreadBy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageReadReceiptMemberResponseDto"
            }
          }
        },
        "required": [
          "totalRecipients",
          "readCount",
          "unreadCount",
          "readBy",
          "unreadBy"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "eventSequence": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "type": {
            "enum": [
              "text",
              "image",
              "file",
              "system"
            ],
            "type": "string"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "nullable": true
          },
          "editedAt": {
            "type": "string",
            "nullable": true
          },
          "mentionedUserIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "replyTo": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageReplyResponseDto"
              }
            ]
          },
          "attachment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageAttachmentResponseDto"
              }
            ]
          },
          "readReceipt": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageReadReceiptResponseDto"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/MessageSenderResponseDto"
          }
        },
        "required": [
          "id",
          "eventSequence",
          "groupId",
          "senderId",
          "type",
          "text",
          "revokedAt",
          "editedAt",
          "mentionedUserIds",
          "replyTo",
          "attachment",
          "readReceipt",
          "createdAt",
          "sender"
        ]
      },
      "MessageListResponseDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponseDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "groupId",
          "items"
        ]
      },
      "MessageAttachmentDto": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string"
          }
        },
        "required": [
          "fileId"
        ]
      },
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "file"
            ]
          },
          "clientMessageId": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "maxLength": 10000
          },
          "attachment": {
            "$ref": "#/components/schemas/MessageAttachmentDto"
          },
          "replyToMessageId": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "EditMessageDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000
          }
        },
        "required": [
          "text"
        ]
      },
...
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "MEMBER",
              "CLI_BOT",
              "ADMIN"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "ResetUserPasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "pattern": "(?=.*[A-Z])"
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
    }
  }
}